requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
fairscale==0.4.4
filelock==3.6.0
fonttools==4.29.1
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pycocoevalcap==1.2
pycocotools==2.0.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
six==1.16.0
timm==0.4.12
tokenizers==0.10.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.15.0
typing-extensions==4.1.1
urllib3==1.26.8
